PAINT-236: Reset language after rtl espresso tests * Call `updateLocale` with correct context * Moved `RtlUiTestUtils` to util package * Created new `RtlActivityTestRule` to automate teardown * Simplified tests
