+++ Paintroid/src/androidTest/java/org/catrobat/paintroid/test/espresso/util/EspressoUtils.java
@@ -20,18 +20,16 @@
 package org.catrobat.paintroid.test.espresso.util;
 
 import android.app.Activity;
-import android.app.Instrumentation;
 import android.content.Context;
+import android.content.res.Configuration;
 import android.content.res.Resources;
 import android.graphics.Bitmap;
 import android.graphics.Paint;
 import android.graphics.Point;
 import android.graphics.PointF;
-import android.os.SystemClock;
 import android.support.test.InstrumentationRegistry;
 import android.support.test.espresso.ViewInteraction;
 import android.support.test.espresso.contrib.DrawerActions;
-import android.view.MotionEvent;
 import android.view.View;
 import android.widget.TableLayout;
 import android.widget.TableRow;
@@ -387,5 +385,9 @@ public final class EspressoUtils {
 
 	public static Resources getResources() {
 		return InstrumentationRegistry.getTargetContext().getResources();
+	}
+
+	public static Configuration getConfiguration() {
+		return getResources().getConfiguration();
 	}
 }
